Experiment running this on SailfishOS emulator :D
It works \o/

I just installed Rust using:

 $ curl -sSf https://static.rust-lang.org/rustup.sh | sh

And then made some modifications to the source code, all of which are
included in this patch, and ran:

 $ cargo build --verbose
 $ cargo run --example sailfishos_factorial --verbose

I also had to install a few packages, such as cmake, but basically that
was it.

Of course, it's just a proof of concept, it's not ready to be shipped.
It breaks for non SailfishOS apps.